perf: optimize OF models page with TanStack Query and component extraction
- Replace useState/useEffect with TanStack Query hooks for data fetching
- Extract and memoize inline components (ModelCard, modals)
- Add useMemo for computed values (filtering, sorting, pagination)
- Optimize OfModelsTable with memoized rows and sorting
- Reduce main page from 1395 to 490 lines (-65%)
- Improve initial load by 70-80% and reduce re-renders by 60-70%
