+++ components/of-models/EditModelModal.tsx
@@ -1,0 +1,215 @@
+'use client';
+
+import React, { useState, useEffect } from 'react';
+import { createPortal } from 'react-dom';
+import { X, Pencil } from 'lucide-react';
+import { toast } from 'sonner';
+import { useUpdateOfModel, type OfModel } from '@/lib/hooks/useOfModels.query';
+
+const statusConfig: Record<string, { bg: string; text: string; dot: string }> = {
+  ACTIVE: {
+    bg: 'bg-emerald-500/10',
+    text: 'text-emerald-400',
+    dot: 'bg-emerald-400',
+  },
+  INACTIVE: {
+    bg: 'bg-zinc-500/10',
+    text: 'text-zinc-400',
+    dot: 'bg-zinc-400',
+  },
+  PENDING: {
+    bg: 'bg-amber-500/10',
+    text: 'text-amber-400',
+    dot: 'bg-amber-400',
+  },
+  ARCHIVED: {
+    bg: 'bg-rose-500/10',
+    text: 'text-rose-400',
+    dot: 'bg-rose-400',
+  },
+};
+
+interface EditModelModalProps {
+  model: OfModel;
+  onClose: () => void;
+  onSuccess: () => void;
+}
+
+export const EditModelModal = React.memo(function EditModelModal({
+  model,
+  onClose,
+  onSuccess,
+}: EditModelModalProps) {
+  const [name, setName] = useState(model.name);
+  const [displayName, setDisplayName] = useState(model.displayName);
+  const [slug, setSlug] = useState(model.slug);
+  const [bio, setBio] = useState(model.bio || '');
+  const [status, setStatus] = useState(model.status);
+  const [mounted, setMounted] = useState(false);
+
+  const updateMutation = useUpdateOfModel();
+
+  useEffect(() => {
+    setMounted(true);
+    return () => setMounted(false);
+  }, []);
+
+  if (!mounted) return null;
+
+  const handleSubmit = async (e: React.FormEvent) => {
+    e.preventDefault();
+
+    if (!name.trim() || !displayName.trim() || !slug.trim()) {
+      toast.error('Please fill in all required fields');
+      return;
+    }
+
+    try {
+      await updateMutation.mutateAsync({
+        id: model.id,
+        name: name.trim(),
+        displayName: displayName.trim(),
+        slug: slug.trim(),
+        bio: bio.trim() || null,
+        status,
+      });
+
+      toast.success('Model updated successfully');
+      onSuccess();
+    } catch (error: any) {
+      toast.error(error.message || 'Failed to update model');
+    }
+  };
+
+  return createPortal(
+    <div className="fixed inset-0 z-50 flex items-center justify-center p-4">
+      <div
+        className="absolute inset-0 bg-black/80 backdrop-blur-sm"
+        onClick={onClose}
+      />
+
+      <div className="relative w-full max-w-lg bg-zinc-900 border border-zinc-800 rounded-2xl shadow-2xl animate-fadeIn overflow-hidden">
+        <div className="absolute top-0 left-1/2 -translate-x-1/2 w-1/2 h-px bg-gradient-to-r from-transparent via-violet-500 to-transparent" />
+
+        <div className="p-6 border-b border-zinc-800">
+          <div className="flex items-center justify-between">
+            <div className="flex items-center gap-3">
+              <div className="p-2 rounded-lg bg-violet-500/10">
+                <Pencil className="w-5 h-5 text-violet-400" />
+              </div>
+              <h2 className="text-xl font-medium text-white">Edit Model</h2>
+            </div>
+            <button
+              onClick={onClose}
+              className="p-2 rounded-lg text-zinc-500 hover:text-white hover:bg-zinc-800 transition-colors"
+            >
+              <X className="w-5 h-5" />
+            </button>
+          </div>
+        </div>
+
+        <form onSubmit={handleSubmit} className="p-6 space-y-5">
+          <div className="grid grid-cols-2 gap-4">
+            <div className="col-span-2 sm:col-span-1">
+              <label className="block text-sm font-medium text-zinc-400 mb-2">
+                Internal Name <span className="text-rose-400">*</span>
+              </label>
+              <input
+                type="text"
+                value={name}
+                onChange={(e) => setName(e.target.value)}
+                className="w-full px-4 py-3 bg-zinc-800/50 border border-zinc-700/50 rounded-xl text-white placeholder-zinc-600 focus:outline-none focus:border-violet-500/50 transition-colors"
+                required
+              />
+            </div>
+
+            <div className="col-span-2 sm:col-span-1">
+              <label className="block text-sm font-medium text-zinc-400 mb-2">
+                Display Name <span className="text-rose-400">*</span>
+              </label>
+              <input
+                type="text"
+                value={displayName}
+                onChange={(e) => setDisplayName(e.target.value)}
+                className="w-full px-4 py-3 bg-zinc-800/50 border border-zinc-700/50 rounded-xl text-white placeholder-zinc-600 focus:outline-none focus:border-violet-500/50 transition-colors"
+                required
+              />
+            </div>
+          </div>
+
+          <div>
+            <label className="block text-sm font-medium text-zinc-400 mb-2">
+              URL Slug <span className="text-rose-400">*</span>
+            </label>
+            <div className="flex items-center">
+              <span className="px-4 py-3 bg-zinc-800 border border-r-0 border-zinc-700/50 rounded-l-xl text-zinc-500 text-sm">
+                /of-models/
+              </span>
+              <input
+                type="text"
+                value={slug}
+                onChange={(e) => setSlug(e.target.value.toLowerCase().replace(/[^a-z0-9-]/g, ''))}
+                className="flex-1 px-4 py-3 bg-zinc-800/50 border border-zinc-700/50 rounded-r-xl text-white placeholder-zinc-600 focus:outline-none focus:border-violet-500/50 transition-colors"
+                required
+              />
+            </div>
+          </div>
+
+          <div>
+            <label className="block text-sm font-medium text-zinc-400 mb-2">
+              Bio
+            </label>
+            <textarea
+              value={bio}
+              onChange={(e) => setBio(e.target.value)}
+              rows={3}
+              className="w-full px-4 py-3 bg-zinc-800/50 border border-zinc-700/50 rounded-xl text-white placeholder-zinc-600 focus:outline-none focus:border-violet-500/50 transition-colors resize-none"
+            />
+          </div>
+
+          <div>
+            <label className="block text-sm font-medium text-zinc-400 mb-2">
+              Status
+            </label>
+            <div className="flex flex-wrap gap-2">
+              {['ACTIVE', 'INACTIVE', 'PENDING', 'ARCHIVED'].map((s) => (
+                <button
+                  key={s}
+                  type="button"
+                  onClick={() => setStatus(s as OfModel['status'])}
+                  className={`flex items-center gap-2 px-4 py-2 rounded-lg text-sm font-medium transition-all ${
+                    status === s
+                      ? `${statusConfig[s].bg} ${statusConfig[s].text} border border-current`
+                      : 'bg-zinc-800/50 border border-zinc-700/50 text-zinc-500 hover:text-zinc-400'
+                  }`}
+                >
+                  <span className={`w-2 h-2 rounded-full ${statusConfig[s].dot}`} />
+                  {s.charAt(0) + s.slice(1).toLowerCase()}
+                </button>
+              ))}
+            </div>
+          </div>
+
+          <div className="flex gap-3 pt-4">
+            <button
+              type="button"
+              onClick={onClose}
+              className="flex-1 px-4 py-3 bg-zinc-800 border border-zinc-700 text-zinc-300 rounded-xl hover:bg-zinc-700 transition-colors"
+            >
+              Cancel
+            </button>
+            <button
+              type="submit"
+              disabled={updateMutation.isPending}
+              className="flex-1 relative px-4 py-3 rounded-xl font-medium text-white overflow-hidden disabled:opacity-50 disabled:cursor-not-allowed"
+            >
+              <div className="absolute inset-0 bg-gradient-to-r from-violet-600 to-fuchsia-600" />
+              <span className="relative">{updateMutation.isPending ? 'Saving...' : 'Save Changes'}</span>
+            </button>
+          </div>
+        </form>
+      </div>
+    </div>,
+    document.body
+  );
+});
