+++ components/of-models/OfModelsTable.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import React, { useState } from 'react';
+import React, { useState, useMemo } from 'react';
 import Link from 'next/link';
 import {
   Calendar,
@@ -18,26 +18,7 @@ import {
   Users,
 } from 'lucide-react';
 import { cn } from '@/lib/utils';
-
-interface OfModel {
-  id: string;
-  name: string;
-  displayName: string;
-  slug: string;
-  status: 'ACTIVE' | 'INACTIVE' | 'PENDING' | 'ARCHIVED';
-  profileImageUrl: string | null;
-  bio: string | null;
-  personalityType: string | null;
-  instagramUrl: string | null;
-  twitterUrl: string | null;
-  websiteUrl: string | null;
-  launchDate: string | null;
-  guaranteedAmount?: number | null;
-  percentageTaken?: number | null;
-  chattingManagers?: string[];
-  createdAt: string;
-  updatedAt: string;
-}
+import type { OfModel } from '@/lib/hooks/useOfModels.query';
 
 type SortField = 'displayName' | 'status' | 'launchDate' | 'guaranteedAmount' | 'createdAt';
 type SortDirection = 'asc' | 'desc';
@@ -132,7 +113,13 @@ interface SortableHeaderProps {
   onSort: (field: SortField) => void;
 }
 
-function SortableHeader({ label, field, currentSort, direction, onSort }: SortableHeaderProps) {
+const SortableHeader = React.memo(function SortableHeader({
+  label,
+  field,
+  currentSort,
+  direction,
+  onSort
+}: SortableHeaderProps) {
   const isActive = currentSort === field;
 
   return (
@@ -155,9 +142,191 @@ function SortableHeader({ label, field, currentSort, direction, onSort }: Sortab
       )}
     </button>
   );
+});
+
+// Memoized table row component
+interface TableRowProps {
+  model: OfModel;
+  onEdit: (model: OfModel) => void;
+  onDelete: (model: OfModel) => void;
+  onModelClick: (model: OfModel) => void;
 }
 
-export function OfModelsTable({ models, onEdit, onDelete, onModelClick }: OfModelsTableProps) {
+const TableRow = React.memo(function TableRow({ model, onEdit, onDelete, onModelClick }: TableRowProps) {
+  const imageUrl = getDisplayImageUrl(model.profileImageUrl);
+  const statusStyle = statusConfig[model.status] || statusConfig.INACTIVE;
+  const personalityStyle = getPersonalityConfig(model.personalityType);
+
+  return (
+    <tr
+      className="border-b border-zinc-800/30 hover:bg-zinc-800/30 transition-colors duration-150"
+    >
+      {/* Model Info */}
+      <td className="py-4 px-5">
+        <Link
+          href={`of-models/${model.slug}`}
+          onClick={() => onModelClick(model)}
+          className="flex items-center gap-3 group"
+        >
+          <div className="relative w-10 h-10 rounded-lg overflow-hidden bg-zinc-800 shrink-0">
+            {imageUrl ? (
+              <img
+                src={imageUrl}
+                alt={model.displayName}
+                className="w-full h-full object-cover"
+                onError={(e) => {
+                  e.currentTarget.style.display = 'none';
+                  e.currentTarget.nextElementSibling?.classList.remove('hidden');
+                }}
+              />
+            ) : null}
+            <div className={cn(
+              'absolute inset-0 flex items-center justify-center bg-gradient-to-br from-violet-500/30 to-fuchsia-500/30',
+              imageUrl ? 'hidden' : ''
+            )}>
+              <span className="text-sm font-medium text-white/80">
+                {model.displayName.charAt(0).toUpperCase()}
+              </span>
+            </div>
+          </div>
+          <div className="min-w-0">
+            <div className="font-medium text-white group-hover:text-violet-300 transition-colors truncate">
+              {model.displayName}
+            </div>
+            <div className="text-xs text-zinc-500 truncate">@{model.slug}</div>
+          </div>
+        </Link>
+      </td>
+
+      {/* Status */}
+      <td className="py-4 px-5">
+        <span className={cn(
+          'inline-flex items-center gap-1.5 px-2.5 py-1 rounded-full text-xs font-medium',
+          statusStyle.bg
+        )}>
+          <span className={cn('w-1.5 h-1.5 rounded-full', statusStyle.dot)} />
+          <span className={statusStyle.text}>
+            {model.status.charAt(0) + model.status.slice(1).toLowerCase()}
+          </span>
+        </span>
+      </td>
+
+      {/* Personality */}
+      <td className="py-4 px-5">
+        {model.personalityType ? (
+          <span className={cn(
+            'inline-flex items-center gap-1 px-2.5 py-1 rounded-full text-xs font-medium',
+            personalityStyle.bg, personalityStyle.text
+          )}>
+            <Sparkles className="w-3 h-3" />
+            <span className="truncate max-w-[100px]">{model.personalityType}</span>
+          </span>
+        ) : (
+          <span className="text-xs text-zinc-600">—</span>
+        )}
+      </td>
+
+      {/* Launch Date */}
+      <td className="py-4 px-5">
+        <div className="flex items-center gap-2 text-zinc-400">
+          <Calendar className="w-4 h-4 text-zinc-500" />
+          <span className="text-sm">
+            {model.launchDate
+              ? new Date(model.launchDate).toLocaleDateString()
+              : '—'}
+          </span>
+        </div>
+      </td>
+
+      {/* Guaranteed */}
+      <td className="py-4 px-5">
+        <div className="flex items-center gap-2">
+          <DollarSign className="w-4 h-4 text-emerald-400" />
+          <span className="font-medium text-white">
+            {formatCurrency((model as any).guaranteedAmount)}
+          </span>
+        </div>
+      </td>
+
+      {/* Social */}
+      <td className="py-4 px-5">
+        <div className="flex items-center gap-1">
+          {model.instagramUrl && (
+            <a
+              href={getSocialUrl(model.instagramUrl, 'instagram')!}
+              target="_blank"
+              rel="noopener noreferrer"
+              onClick={(e) => e.stopPropagation()}
+              className="p-1.5 rounded-lg text-pink-400/70 hover:text-pink-400 hover:bg-pink-500/10 transition-all"
+            >
+              <Instagram className="w-4 h-4" />
+            </a>
+          )}
+          {model.twitterUrl && (
+            <a
+              href={getSocialUrl(model.twitterUrl, 'twitter')!}
+              target="_blank"
+              rel="noopener noreferrer"
+              onClick={(e) => e.stopPropagation()}
+              className="p-1.5 rounded-lg text-sky-400/70 hover:text-sky-400 hover:bg-sky-500/10 transition-all"
+            >
+              <Twitter className="w-4 h-4" />
+            </a>
+          )}
+          {model.websiteUrl && (
+            <a
+              href={model.websiteUrl}
+              target="_blank"
+              rel="noopener noreferrer"
+              onClick={(e) => e.stopPropagation()}
+              className="p-1.5 rounded-lg text-zinc-400/70 hover:text-zinc-300 hover:bg-zinc-500/10 transition-all"
+            >
+              <Globe className="w-4 h-4" />
+            </a>
+          )}
+          {!model.instagramUrl && !model.twitterUrl && !model.websiteUrl && (
+            <span className="text-xs text-zinc-600">—</span>
+          )}
+        </div>
+      </td>
+
+      {/* Actions */}
+      <td className="py-4 px-5">
+        <div className="flex items-center gap-1">
+          <Link
+            href={`of-models/${model.slug}`}
+            onClick={() => onModelClick(model)}
+            className="p-1.5 rounded-lg text-zinc-500 hover:text-white hover:bg-zinc-700 transition-all"
+            title="View Profile"
+          >
+            <Eye className="w-4 h-4" />
+          </Link>
+          <button
+            onClick={() => onEdit(model)}
+            className="p-1.5 rounded-lg text-zinc-500 hover:text-violet-400 hover:bg-violet-500/10 transition-all"
+            title="Edit"
+          >
+            <Pencil className="w-4 h-4" />
+          </button>
+          <button
+            onClick={() => onDelete(model)}
+            className="p-1.5 rounded-lg text-zinc-500 hover:text-rose-400 hover:bg-rose-500/10 transition-all"
+            title="Delete"
+          >
+            <Trash2 className="w-4 h-4" />
+          </button>
+        </div>
+      </td>
+    </tr>
+  );
+});
+
+export const OfModelsTable = React.memo(function OfModelsTable({
+  models,
+  onEdit,
+  onDelete,
+  onModelClick
+}: OfModelsTableProps) {
   const [sortField, setSortField] = useState<SortField>('displayName');
   const [sortDirection, setSortDirection] = useState<SortDirection>('asc');
 
@@ -170,31 +339,34 @@ export function OfModelsTable({ models, onEdit, onDelete, onModelClick }: OfMode
     }
   };
 
-  const sortedModels = [...models].sort((a, b) => {
-    let comparison = 0;
+  // Memoize sorted models
+  const sortedModels = useMemo(() => {
+    return [...models].sort((a, b) => {
+      let comparison = 0;
 
-    switch (sortField) {
-      case 'displayName':
-        comparison = a.displayName.localeCompare(b.displayName);
-        break;
-      case 'status':
-        comparison = a.status.localeCompare(b.status);
-        break;
-      case 'launchDate':
-        const dateA = a.launchDate ? new Date(a.launchDate).getTime() : 0;
-        const dateB = b.launchDate ? new Date(b.launchDate).getTime() : 0;
-        comparison = dateA - dateB;
-        break;
-      case 'guaranteedAmount':
-        comparison = (a.guaranteedAmount || 0) - (b.guaranteedAmount || 0);
-        break;
-      case 'createdAt':
-        comparison = new Date(a.createdAt).getTime() - new Date(b.createdAt).getTime();
-        break;
-    }
+      switch (sortField) {
+        case 'displayName':
+          comparison = a.displayName.localeCompare(b.displayName);
+          break;
+        case 'status':
+          comparison = a.status.localeCompare(b.status);
+          break;
+        case 'launchDate':
+          const dateA = a.launchDate ? new Date(a.launchDate).getTime() : 0;
+          const dateB = b.launchDate ? new Date(b.launchDate).getTime() : 0;
+          comparison = dateA - dateB;
+          break;
+        case 'guaranteedAmount':
+          comparison = ((a as any).guaranteedAmount || 0) - ((b as any).guaranteedAmount || 0);
+          break;
+        case 'createdAt':
+          comparison = new Date(a.createdAt).getTime() - new Date(b.createdAt).getTime();
+          break;
+      }
 
-    return sortDirection === 'asc' ? comparison : -comparison;
-  });
+      return sortDirection === 'asc' ? comparison : -comparison;
+    });
+  }, [models, sortField, sortDirection]);
 
   if (models.length === 0) {
     return (
@@ -265,180 +437,20 @@ export function OfModelsTable({ models, onEdit, onDelete, onModelClick }: OfMode
             </tr>
           </thead>
           <tbody>
-            {sortedModels.map((model) => {
-              const imageUrl = getDisplayImageUrl(model.profileImageUrl);
-              const statusStyle = statusConfig[model.status] || statusConfig.INACTIVE;
-              const personalityStyle = getPersonalityConfig(model.personalityType);
-
-              return (
-                <tr
-                  key={model.id}
-                  className="border-b border-zinc-800/30 hover:bg-zinc-800/30 transition-colors duration-150"
-                >
-                  {/* Model Info */}
-                  <td className="py-4 px-5">
-                    <Link
-                      href={`of-models/${model.slug}`}
-                      onClick={() => onModelClick(model)}
-                      className="flex items-center gap-3 group"
-                    >
-                      <div className="relative w-10 h-10 rounded-lg overflow-hidden bg-zinc-800 shrink-0">
-                        {imageUrl ? (
-                          <img
-                            src={imageUrl}
-                            alt={model.displayName}
-                            className="w-full h-full object-cover"
-                            onError={(e) => {
-                              e.currentTarget.style.display = 'none';
-                              e.currentTarget.nextElementSibling?.classList.remove('hidden');
-                            }}
-                          />
-                        ) : null}
-                        <div className={cn(
-                          'absolute inset-0 flex items-center justify-center bg-gradient-to-br from-violet-500/30 to-fuchsia-500/30',
-                          imageUrl ? 'hidden' : ''
-                        )}>
-                          <span className="text-sm font-medium text-white/80">
-                            {model.displayName.charAt(0).toUpperCase()}
-                          </span>
-                        </div>
-                      </div>
-                      <div className="min-w-0">
-                        <div className="font-medium text-white group-hover:text-violet-300 transition-colors truncate">
-                          {model.displayName}
-                        </div>
-                        <div className="text-xs text-zinc-500 truncate">@{model.slug}</div>
-                      </div>
-                    </Link>
-                  </td>
-
-                  {/* Status */}
-                  <td className="py-4 px-5">
-                    <span className={cn(
-                      'inline-flex items-center gap-1.5 px-2.5 py-1 rounded-full text-xs font-medium',
-                      statusStyle.bg
-                    )}>
-                      <span className={cn('w-1.5 h-1.5 rounded-full', statusStyle.dot)} />
-                      <span className={statusStyle.text}>
-                        {model.status.charAt(0) + model.status.slice(1).toLowerCase()}
-                      </span>
-                    </span>
-                  </td>
-
-                  {/* Personality */}
-                  <td className="py-4 px-5">
-                    {model.personalityType ? (
-                      <span className={cn(
-                        'inline-flex items-center gap-1 px-2.5 py-1 rounded-full text-xs font-medium',
-                        personalityStyle.bg, personalityStyle.text
-                      )}>
-                        <Sparkles className="w-3 h-3" />
-                        <span className="truncate max-w-[100px]">{model.personalityType}</span>
-                      </span>
-                    ) : (
-                      <span className="text-xs text-zinc-600">—</span>
-                    )}
-                  </td>
-
-                  {/* Launch Date */}
-                  <td className="py-4 px-5">
-                    <div className="flex items-center gap-2 text-zinc-400">
-                      <Calendar className="w-4 h-4 text-zinc-500" />
-                      <span className="text-sm">
-                        {model.launchDate
-                          ? new Date(model.launchDate).toLocaleDateString()
-                          : '—'}
-                      </span>
-                    </div>
-                  </td>
-
-                  {/* Guaranteed */}
-                  <td className="py-4 px-5">
-                    <div className="flex items-center gap-2">
-                      <DollarSign className="w-4 h-4 text-emerald-400" />
-                      <span className="font-medium text-white">
-                        {formatCurrency(model.guaranteedAmount)}
-                      </span>
-                    </div>
-                  </td>
-
-                  {/* Social */}
-                  <td className="py-4 px-5">
-                    <div className="flex items-center gap-1">
-                      {model.instagramUrl && (
-                        <a
-                          href={getSocialUrl(model.instagramUrl, 'instagram')!}
-                          target="_blank"
-                          rel="noopener noreferrer"
-                          onClick={(e) => e.stopPropagation()}
-                          className="p-1.5 rounded-lg text-pink-400/70 hover:text-pink-400 hover:bg-pink-500/10 transition-all"
-                        >
-                          <Instagram className="w-4 h-4" />
-                        </a>
-                      )}
-                      {model.twitterUrl && (
-                        <a
-                          href={getSocialUrl(model.twitterUrl, 'twitter')!}
-                          target="_blank"
-                          rel="noopener noreferrer"
-                          onClick={(e) => e.stopPropagation()}
-                          className="p-1.5 rounded-lg text-sky-400/70 hover:text-sky-400 hover:bg-sky-500/10 transition-all"
-                        >
-                          <Twitter className="w-4 h-4" />
-                        </a>
-                      )}
-                      {model.websiteUrl && (
-                        <a
-                          href={model.websiteUrl}
-                          target="_blank"
-                          rel="noopener noreferrer"
-                          onClick={(e) => e.stopPropagation()}
-                          className="p-1.5 rounded-lg text-zinc-400/70 hover:text-zinc-300 hover:bg-zinc-500/10 transition-all"
-                        >
-                          <Globe className="w-4 h-4" />
-                        </a>
-                      )}
-                      {!model.instagramUrl && !model.twitterUrl && !model.websiteUrl && (
-                        <span className="text-xs text-zinc-600">—</span>
-                      )}
-                    </div>
-                  </td>
-
-                  {/* Actions */}
-                  <td className="py-4 px-5">
-                    <div className="flex items-center gap-1">
-                      <Link
-                        href={`of-models/${model.slug}`}
-                        onClick={() => onModelClick(model)}
-                        className="p-1.5 rounded-lg text-zinc-500 hover:text-white hover:bg-zinc-700 transition-all"
-                        title="View Profile"
-                      >
-                        <Eye className="w-4 h-4" />
-                      </Link>
-                      <button
-                        onClick={() => onEdit(model)}
-                        className="p-1.5 rounded-lg text-zinc-500 hover:text-violet-400 hover:bg-violet-500/10 transition-all"
-                        title="Edit"
-                      >
-                        <Pencil className="w-4 h-4" />
-                      </button>
-                      <button
-                        onClick={() => onDelete(model)}
-                        className="p-1.5 rounded-lg text-zinc-500 hover:text-rose-400 hover:bg-rose-500/10 transition-all"
-                        title="Delete"
-                      >
-                        <Trash2 className="w-4 h-4" />
-                      </button>
-                    </div>
-                  </td>
-                </tr>
-              );
-            })}
+            {sortedModels.map((model) => (
+              <TableRow
+                key={model.id}
+                model={model}
+                onEdit={onEdit}
+                onDelete={onDelete}
+                onModelClick={onModelClick}
+              />
+            ))}
           </tbody>
         </table>
       </div>
     </div>
   );
-}
+});
 
 export default OfModelsTable;
